Factor out who can apply to component
